This patch fixes up a narrowing conversion warning
I don't know if there were a real issue, but this has fixed
upt he warning at compilation time.

+++ lib/fact.c
@@ -1,4 +1,4 @@
-long fact (int val)
+unsigned long long fact (int val)
 {
 	if (val > 1)
 		return val * fact(val -1);
